File structure changes and moving code around(2).
