+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drivetrain/MechDrive.java
@@ -1,7 +1,5 @@
 package org.firstinspires.ftc.teamcode.drivetrain;
 
-import com.acmerobotics.roadrunner.ftc.OverflowEncoder;
-import com.acmerobotics.roadrunner.ftc.RawEncoder;
 import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
@@ -10,7 +8,6 @@ import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.hardware.IMU;
 
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
-import org.firstinspires.ftc.teamcode.teleop.MecanumDrive;
 
 public class MechDrive {
     public final double MAX_POWER = 0.9;
